Add AvatarService and refactor
