+++ Sources/Gravatar/Network/CancellableDataTask.swift
@@ -2,12 +2,13 @@ import Foundation
 
 /// Represents the task of a data downloading process, which can be cancelled while it's running.
 ///
-/// We offer a default implementation of `CancellableDataTask` for `URLSessionTask` and `Task`.
+/// We offer a default implementation of `CancellableDataTask` for `Task`.
 public protocol CancellableDataTask {
     /// Cancells a running task.
     func cancel()
-}
 
-extension URLSessionTask: CancellableDataTask {}
+    /// Was the task cancelled?
+    var isCancelled: Bool { get }
+}
 
 extension Task: CancellableDataTask {}
